Split up TopologyDisplayView and remove certain state creation for empty layout
